fix(composition-runtime): share reversed audio buffer helper and cover after-extension for reversed clips
Extract createReversedAudioBuffer to a shared util (using AudioBuffer
constructor instead of OfflineAudioContext) and decrease trimBefore by the
after-extension for reversed clips so transition tails decode the source
window they actually request.
